# openapi.json
{
  "components": {
    "schemas": {
...
      "ChatStreamEvent": {
        "discriminator": {
          "mapping": {
            "done": "#/components/schemas/DoneEvent",
            "error": "#/components/schemas/ErrorEvent",
            "text-delta": "#/components/schemas/TextDeltaEvent",
            "tool-call-begin": "#/components/schemas/ToolCallBeginEvent",
            "tool-call-delta": "#/components/schemas/ToolCallDeltaEvent"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/TextDeltaEvent"
          },
          {
            "$ref": "#/components/schemas/ToolCallBeginEvent"
          },
          {
            "$ref": "#/components/schemas/ToolCallDeltaEvent"
          },
          {
            "$ref": "#/components/schemas/DoneEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      },
      "ChatStreamMessage": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatStreamMessage",
        "type": "object"
      },
      "ChatStreamRequest": {
        "example": {
          "messages": [
            {
              "content": "Hello, how are you?",
              "role": "user"
            }
          ]
        },
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatStreamMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "run_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Config"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "required": [
          "messages"
        ],
        "title": "ChatStreamRequest",
        "type": "object"
      },
...
      "CreateProjectRequest": {
        "additionalProperties": false,
        "example": {
          "name": "New Project"
        },
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "DoneEvent": {
        "properties": {
          "finish_reason": {
            "enum": [
              "stop",
              "length",
              "error"
            ],
            "title": "Finish Reason",
            "type": "string"
          },
          "type": {
            "const": "done",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "finish_reason"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorEvent": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Code"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "const": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
...
      "PaginatedProjects": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "default": false,
            "title": "Hasmore",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedProjects",
        "type": "object"
      },
      "Project": {
        "additionalProperties": false,
        "example": {
          "chatCount": 2,
          "createdAt": "2024-01-15T10:00:00Z",
          "id": "work",
          "name": "Work"
        },
        "properties": {
          "chatCount": {
            "title": "Chatcount",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "chatCount"
        ],
        "title": "Project",
        "type": "object"
      },
      "TextDeltaEvent": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "type": {
            "const": "text-delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "delta"
        ],
        "title": "TextDeltaEvent",
        "type": "object"
      },
      "ToolCallBeginEvent": {
        "properties": {
          "tool_call_id": {
            "title": "Tool Call Id",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "type": {
            "const": "tool-call-begin",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "tool_call_id",
          "tool_name"
        ],
        "title": "ToolCallBeginEvent",
        "type": "object"
      },
      "ToolCallDeltaEvent": {
        "properties": {
          "args_delta": {
            "title": "Args Delta",
            "type": "string"
          },
          "tool_call_id": {
            "title": "Tool Call Id",
            "type": "string"
          },
          "type": {
            "const": "tool-call-delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "tool_call_id",
          "args_delta"
        ],
        "title": "ToolCallDeltaEvent",
        "type": "object"
      },
...
      "UpdateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateProjectRequest",
        "type": "object"
      },
      "UserSettings": {
        "additionalProperties": false,
        "example": {
          "email": "john@example.com",
          "name": "John Doe",
          "notifications": true
        },
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notifications": {
            "title": "Notifications",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "email",
          "notifications"
        ],
        "title": "UserSettings",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Databricks Apps FastAPI Starter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat/stream": {
      "post": {
        "operationId": "chatStream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStreamEvent"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/projects": {
      "get": {
        "operationId": "listProjects",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjects"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{projectId}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "updateProject",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "getUserSettings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get User Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "updateUserSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Update User Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}